+++ rpcprueba/calculadoraResp/suma_server.c
@@ -5,11 +5,11 @@
  */
 
 #include "suma.h"
+#include <math.h>
 
-float *
-suma_1_svc(sumandos *argp, struct svc_req *rqstp)
+int * suma_1_svc(sumandos *argp, struct svc_req *rqstp)
 {
-	static float  result;
+	static int  result;
 
 	/*
 	 * insert server code here
@@ -23,15 +23,14 @@ suma_1_svc(sumandos *argp, struct svc_req *rqstp)
 		break;
 		case 4: result=argp->sumando1 * argp->sumando2;
 		break;
-		case 5: result=(sin(argp->sumando1));
+		case 5: result=(int)(sin(argp->sumando1));
 		break;
-		case 6: result=(cos(argp->sumando1));
+		case 6: result=(int)(cos(argp->sumando1));
 		break;
-		case 7: result=(tan(argp->sumando1));
+		case 7: result=(int)(tan(argp->sumando1));
 		break;
 		default: break;
 	
 	}
- 
 	return &result;
 }
